+++ src/visual.tsx
@@ -161,7 +161,8 @@ export class Visual implements IVisual {
         }
         batch_requests(
             input_properties,
-            this.formattingSettings.advanced_settings.offset_multiplier.value
+            this.formattingSettings.advanced_settings.offset_multiplier.value,
+            this.formattingSettings.advanced_settings.backend_url.value
         ).then(
             (returned_features)=>{
                 const non_mappable_rows:NonMappableRow[] = []
@@ -246,7 +247,7 @@ export class Visual implements IVisual {
         this.react_root.render(
             <NickMap
                 host={this.host}
-                version_node                          = {<>v4.4.0<br/>NickMapBI</>}
+                version_node                          = {<>v4.4.2.0<br/>NickMapBI</>}
 
                 layer_arcgis_rest_url                 = {map_background_settings.url_tile_arcgis.value}
                 layer_arcgis_rest_show_initial        = {map_background_settings.url_tile_arcgis_show.value}
@@ -256,8 +257,8 @@ export class Visual implements IVisual {
 
                 layer_road_network_show_initial       = {road_network_settings.show.value}
                 layer_road_network_ticks_show_initial = {road_network_settings.show_ticks.value}
-                layer_road_network_state_color       = {road_network_settings.state_road_color.value.value}
-                layer_road_network_psp_color         = {road_network_settings.psp_color.value.value}
+                layer_road_network_state_color        = {road_network_settings.state_road_color.value.value}
+                layer_road_network_psp_color          = {road_network_settings.psp_color.value.value}
 
                 layer_raster_brightness               = {map_background_settings.osm_brightness.value}
                 layer_raster_contrast                 = {map_background_settings.osm_contrast.value}
@@ -271,6 +272,7 @@ export class Visual implements IVisual {
                 show_result_count                     = {map_status_bar_settings.show_result_count.value}
 
                 allow_drag_box_selection              = {advanced_settings.allow_drag_box_selection.value}
+                backend_url                           = {advanced_settings.backend_url.value}
 
                 feature_collection                    = {this.feature_collection}
                 feature_collection_request_count      = {this.features_requested_count}